services/edge-simulator/requirements.txt
paho-mqtt==1.6.1
numpy==1.26.0
scipy==1.11.3
python-dotenv==1.0.0
requests==2.31.0
